refactored executors, and made a new wikitables executor
