+++ allennlp/tests/fixtures/semantic_parsing/nlvr_coverage_semantic_parser/serialization/vocabulary/non_padded_namespaces.txt
@@ -1,2 +1,2 @@
-denotations
 rule_labels
+denotations
